[not upstreamable] mtd: spi-nor: add Global Block Unlock support
Proper fix would be to add support for individual block lock protection
and to use the global block unlock support as an optimization when we
want to unlock all the individual blocks.

We can't determine this purely by manufacturer type and it's not
autodetectable by anything like SFDP, so make a new flag for it:
UNLOCK_GLOBAL_BLOCK.

Note that the Global Block Unlock command has different names
depending on the manufacturer, but always the same command value:
0x98. Macronix's MX25U12835F names it Gang Block Unlock,
Winbound's W25Q128FV names it Global Block Unlock and
Microchip's SST26VF064B names it Global Block Protection Unlock.

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>
[eugen.hristev@microchip.com: Ported to 2021.01]
Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>
[eugen.hristev@microchip.com: Ported to 2022.01]
Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>
[varshini.rajendran@microchip.com: Ported to 2025.07]
Signed-off-by: Varshini Rajendran <varshini.rajendran@microchip.com>