pinctrl: at91-pio4: Rename drive strength macros
Rename drive strength macros to avoid confusion of different meanings
for different architectures that use this driver.

Instead of using the macros for low, med, high the macros will be
replaced with the value that will be placed in the register.

Signed-off-by: Ryan Wanner <Ryan.Wanner@microchip.com>
(cherry picked from commit 7f15b6eaf8302320c96c161cec021855df7e8095)