clk: at91: clk-generic: Add status check
Add a clock status check for the generated clocks. This ensures the
generic clock is ready after being modified.

Since this only applies for certain SoCs a bool has been added to the
gck_pcr_layout struct so this check will only be triggered for SoCs that
support this read.

Signed-off-by: Ryan Wanner <Ryan.Wanner@microchip.com>