clk: at91: pmc: Add clk enable to clock set up
Add a clock enable parameter to the set up function. This is needed due
to some cases where the PLL needs to be enabled at registration in order
to prevent clocks from being assigned to parents that are not enabled.

Setting clocks to parents that are not enabled can cause undefined
behavior this prevents that case from happening.

Signed-off-by: Ryan Wanner <Ryan.Wanner@microchip.com>